fix(store): validate attribute definitions inside the write transaction
Load and lock the definition row within each set/supersede attempt so a
concurrent deactivation or option change committing after an outside-tx
check cannot be bypassed. Also widen PostgreSQL display_order to BIGINT
to match the int64 contract and SQLite's 64-bit integers.

+++ internal/store/schema_pg.sql
@@ -447,7 +447,7 @@ CREATE TABLE IF NOT EXISTS attribute_definitions (
     field_type     TEXT NOT NULL,
     record_target  TEXT,
     cardinality    TEXT NOT NULL DEFAULT 'single',
-    display_order  INTEGER NOT NULL DEFAULT 0,
+    display_order  BIGINT NOT NULL DEFAULT 0,
     is_required    BOOLEAN NOT NULL DEFAULT FALSE,
     ownership      TEXT NOT NULL DEFAULT 'user',
     ui_creatable   BOOLEAN NOT NULL DEFAULT TRUE,
